# openapi.yaml
openapi: 3.0.3
info:
  title: Flex-Restro API
  description: API for Global Restaurant SaaS Platform
  version: 1.0.0
servers:
  - url: https://api.flex-restro.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [owner, staff, admin]
        tenant_id:
          type: string
        created_at:
          type: string
          format: date-time
    Restaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        currency:
          type: string
        tax_rate:
          type: number
        qr_code_url:
          type: string
        tenant_id:
          type: string
        created_at:
          type: string
          format: date-time
    Menu:
      type: object
      properties:
        id:
          type: string
        restaurant_id:
          type: string
        name:
          type: string
        description:
          type: string
        image_url:
          type: string
        available:
          type: boolean
        created_at:
          type: string
          format: date-time
    MenuItem:
      type: object
      properties:
        id:
          type: string
        menu_id:
          type: string
        name:
          type: string
        price:
          type: number
        variants:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              price:
                type: number
        category:
          type: string
        image_url:
          type: string
        available:
          type: boolean
        created_at:
          type: string
          format: date-time
    Order:
      type: object
      properties:
        id:
          type: string
        restaurant_id:
          type: string
        customer_id:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              item_id:
                type: string
              quantity:
                type: integer
              variants:
                type: array
                items:
                  type: string
        total:
          type: number
        status:
          type: string
          enum: [pending, confirmed, preparing, ready, delivered, cancelled]
        payment_id:
          type: string
        created_at:
          type: string
          format: date-time
    Payment:
      type: object
      properties:
        id:
          type: string
        order_id:
          type: string
        amount:
          type: number
        currency:
          type: string
        gateway:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        created_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [owner, staff]
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants:
    post:
      summary: Register a new restaurant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                currency:
                  type: string
                tax_rate:
                  type: number
      responses:
        '201':
          description: Restaurant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get restaurants for user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of restaurants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Restaurant'
  /restaurants/{restaurant_id}:
    get:
      summary: Get restaurant details
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restaurant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update restaurant
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                address:
                  type: string
                currency:
                  type: string
                tax_rate:
                  type: number
      responses:
        '200':
          description: Restaurant updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/menus:
    post:
      summary: Create a new menu
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                image_url:
                  type: string
                available:
                  type: boolean
      responses:
        '201':
          description: Menu created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get menus for restaurant
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of menus
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Menu'
  /restaurants/{restaurant_id}/menus/{menu_id}:
    get:
      summary: Get menu details
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Menu details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update menu
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                image_url:
                  type: string
                available:
                  type: boolean
      responses:
        '200':
          description: Menu updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Menu'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete menu
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Menu deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/menus/{menu_id}/items:
    post:
      summary: Add menu item
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                variants:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      price:
                        type: number
                category:
                  type: string
                image_url:
                  type: string
                available:
                  type: boolean
      responses:
        '201':
          description: Item added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get menu items
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenuItem'
  /restaurants/{restaurant_id}/menus/{menu_id}/items/{item_id}:
    put:
      summary: Update menu item
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: number
                variants:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      price:
                        type: number
                category:
                  type: string
                image_url:
                  type: string
                available:
                  type: boolean
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuItem'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete menu item
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
        - name: menu_id
          in: path
          required: true
          schema:
            type: string
        - name: item_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Item deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /restaurants/{restaurant_id}/orders:
    post:
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_id:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      item_id:
                        type: string
                      quantity:
                        type: integer
                      variants:
                        type: array
                        items:
                          type: string
                payment_method:
                  type: string
                  enum: [card, upi, wallet, cash]
      responses:
        '201':
          description: Order placed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get orders for restaurant
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /restaurants/{restaurant_id}/orders/{order_id}:
    get:
      summary: Get order details
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
        - name: order_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Order details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - name: restaurant_id
          in: path
          required: true
          schema:
            type: string
        - name: order_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, preparing, ready, delivered, cancelled]
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /webhooks/marketplace/{marketplace}:
    post:
      summary: Receive order from marketplace
      parameters:
        - name: marketplace
          in: path
          required: true
          schema:
            type: string
            enum: [swiggy, zomato]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order_id:
                  type: string
                restaurant_id:
                  type: string
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      quantity:
                        type: integer
                      price:
                        type: number
                total:
                  type: number
                customer:
                  type: object
                  properties:
                    name:
                      type: string
                    phone:
                      type: string
      responses:
        '200':
          description: Order received
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  order_id:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /payments/webhook:
    post:
      summary: Payment gateway webhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                payment_id:
                  type: string
                order_id:
                  type: string
                status:
                  type: string
                  enum: [completed, failed]
                amount:
                  type: number
      responses:
        '200':
          description: Webhook processed
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'